Extract order-analysis pipeline
