+++ docs/ai/repo-map.md
@@ -37,7 +37,7 @@ Scope: single source of truth for detailed file layout, entry points, package st
 - `infra/processing/`: signal processing pipeline (buffers, FFT, payload shaping, and processor facade).
 - `infra/config/`: runtime settings store (single `SettingsStore` owns both analysis and device settings) used by runtime wiring and exposed to use-cases through focused shared read ports.
 - `infra/workers/`: worker-pool infrastructure.
-- `use_cases/diagnostics/`: post-stop diagnostics pipeline. See `docs/analysis_pipeline.md` for the full data flow, pipeline steps, and module map. `findings.py` owns reference-check and peak/order orchestration, `peak_binning.py` owns persistent-peak accumulation/scoring, `signal_aggregation.py` owns speed/location summary aggregation, `math_utils.py` owns generic statistics/correlation helpers, `speed_profile_helpers.py` owns speed-profile assembly helpers, and `order_heuristics.py` owns the pure diffuse-excitation, localization-override, and engine-alias heuristics separated from `order_analysis.py`; `top_cause_selection.py` delegates classification and ranking to the domain `Finding`; `location_analysis.py` owns the location-analysis pipeline and `LocationAnalysisResult` typed return; `_types.py` is now limited to diagnostics-local helper types such as `AccelStatistics`, while boundary summary payloads live in `shared/boundaries/analysis_payload.py`. The package `__init__.py` re-exports only high-level analysis entrypoints; PDF rendering types/helpers now live under `adapters/pdf/`.
+- `use_cases/diagnostics/`: post-stop diagnostics pipeline. See `docs/analysis_pipeline.md` for the full data flow, pipeline steps, and module map. `findings.py` owns reference-check and peak/order orchestration, `peak_binning.py` owns persistent-peak accumulation/scoring, `signal_aggregation.py` owns speed/location summary aggregation, `math_utils.py` owns generic statistics/correlation helpers, `speed_profile_helpers.py` owns speed-profile assembly helpers, `order_analysis.py` now owns the core order-matching/scoring/assembly primitives, `order_pipeline.py` owns `OrderAnalysisSession` plus the `_build_order_findings()`/multi-location orchestration layer above those primitives, and `order_heuristics.py` owns the pure diffuse-excitation, localization-override, and engine-alias heuristics; `top_cause_selection.py` delegates classification and ranking to the domain `Finding`; `location_analysis.py` owns the location-analysis pipeline and `LocationAnalysisResult` typed return; `_types.py` is now limited to diagnostics-local helper types such as `AccelStatistics`, while boundary summary payloads live in `shared/boundaries/analysis_payload.py`. The package `__init__.py` re-exports only high-level analysis entrypoints; PDF rendering types/helpers now live under `adapters/pdf/`.
 - `use_cases/history/`: run query/delete, PDF report generation, CSV/ZIP exports, and history-facing helper orchestration above persistence ports. `helpers.py` owns the local `HistoryRecord` TypedDict plus generic history workflow helpers, while `report_interpretation.py` owns pure report-domain interpretation consumed by the PDF mapper.
 - `use_cases/run/`: recording pipeline orchestration; `lifecycle_state.py` owns `RunLifecycleState` plus the active-run snapshot, `persistence_writer.py` owns history-write coordination/retry bookkeeping, `logger.py` owns `RunRecorder`, `post_analysis.py` owns the background analysis queue above the injected run-persistence/error boundary, and `sample_builder.py` owns pure sample-building helpers.
 - `use_cases/updates/`: wheel-based updater workflow orchestration, runtime prerequisite validation, firmware cache, ESP flashing, release discovery, install, rollback, runner, Wi-Fi, and status tracking.
